Use quadletFunctions for
* vaultWarden
* caddy
* home-assistant
* zigbee2mqtt

+++ quadlet/caddyQuadlet.sh
@@ -1,8 +1,18 @@
 #!/bin/bash
 set -euo pipefail
+source /links/bin/lib/quadletFunctions.sh
 
-function CreateQuadletCaddy() {
-  mkdir -p ${CADDY_ROOT_DIR}/data
+function CreateCaddyFile() {
+if [ -f ${CADDYFILE} ]; then
+  echo "CADDYFILE ${CADDYFILE} exists"
+  return
+fi
+if [ ${INTERNAL_TLS} != "true" ]; then
+  TLS_CONFIG="# tls internal"
+else
+  TLS_CONFIG="tls internal"
+fi  
+  mkdir -p $(dirname ${CADDYFILE})
   cat <<EOF > ${CADDYFILE}
 # Caddy file
 # to enable tls again replace :8000 with :80 and restart Caddy
@@ -45,7 +55,7 @@ ${CADDY_PROXY_DOMAIN}:${NEXTCLOUD_HTTPS_PORT} {
   }
   
   # for local testing uncomment the following line
-  tls internal
+  ${TLS_CONFIG}
 
   # Change below to host IP
   reverse_proxy ${SERVER_IP}:${NEXTCLOUD_HTTP_PORT}
@@ -54,13 +64,15 @@ ${CADDY_PROXY_DOMAIN}:${NEXTCLOUD_HTTPS_PORT} {
 ${CADDY_PROXY_DOMAIN}:${VAULTWARDEN_HTTPS_PORT} {
 	reverse_proxy ${SERVER_IP}:${VAULTWARDEN_ROCKET_PORT}
   # for local testing uncomment the following line
-  # tls internal
+  ${TLS_CONFIG}
 }
 EOF
+}
 
-  cat <<EOF > ${QUADLET_DIR}/caddy.container 
+function CreateQuadlet() {
+  cat <<EOF > ${QUADLET_DIR}/${SERVICE_NAME}.container 
 [Unit]
-Description=caddy
+Description=${SERVICE_NAME}
 Wants=network-online.target
 After=network-online.target
 
@@ -68,177 +80,76 @@ After=network-online.target
 # Pod=nextcloud.pod
 Label=app=nextcloud
 AutoUpdate=${PODMAN_AUTO_UPDATE_STRATEGY}
-ContainerName=caddy
-Image=docker.io/caddy:latest
+ContainerName=
+Image=${CONTAINER_IMAGE}
 Network=${NETWORK_NAME}
 # PublishPort=80:80 # required for acme challenge which needs to run every other month
 PublishPort=${NEXTCLOUD_HTTP_PORT}:${NEXTCLOUD_HTTP_PORT}
 PublishPort=${NEXTCLOUD_HTTPS_PORT}:${NEXTCLOUD_HTTPS_PORT}
 PublishPort=${VAULTWARDEN_HTTPS_PORT}:${VAULTWARDEN_HTTPS_PORT}
 Volume=${CADDYFILE}:/etc/caddy/Caddyfile:z
-Volume=${CADDY_ROOT_DIR}/data:/data:Z
-Volume=${NEXTCLOUD_ROOT_DIR}/html:/var/www/html:ro,z
+Volume=${DATA_DIR}/data:/data:Z
+Volume=${NEXTCLOUD_DATA_DIR}/html:/var/www/html:ro,z
 AddCapability=CAP_AUDIT_WRITE
 
 [Install]
-WantedBy=default.target
+${START_ON_BOOT}
 EOF
 }
 
-function CreatePodmanNetwork() {
-  podman network create ${NETWORK_NAME} --ignore
-}
-
-
 function postInstall() {
   ${SYSTEMCTL_CMD} daemon-reload
-  ${SYSTEMCTL_CMD} start caddy.service
+  ${SYSTEMCTL_CMD} start ${SERVICE_NAME}.service
   # for local development enable this
   #  if [[ $(id -u) -eq 0 ]] ; then 
-  #  cp ${CADDY_ROOT_DIR}/data/caddy/pki/authorities/local/root.crt /etc/pki/ca-trust/source/anchors/caddy-root-ca.crt
+  #  cp ${DATA_DIR}/data/caddy/pki/authorities/local/root.crt /etc/pki/ca-trust/source/anchors/caddy-root-ca.crt
   #  update-ca-trust
   # fi  
 }
 
-function uninstallCaddy() {
-  ${SYSTEMCTL_CMD} stop caddy.service
-  rm ${QUADLET_DIR}/caddy.container
+function install() {
+  CreatePodmanNetwork
+  CreateCaddyFile
+  CreateQuadlet
+  postInstall
+  showStatus
 }
 
 function setEnvVars() {
-  if [[ $(id -u) -eq 0 ]] ; then 
-    echo "running as USER root" 
-    export QUADLET_DIR=/etc/containers/systemd
-    export SYSTEMD_UNIT_DIR=/etc/systemd/system
-    export SYSTEMCTL_CMD="systemctl"
-  else  
-    echo "running as USER ${USER}" 
-    export QUADLET_DIR=${HOME}/.config/containers/systemd
-    export SYSTEMD_UNIT_DIR=${HOME}/.config/systemd/user
-    export SYSTEMCTL_CMD="systemctl --user"
-  fi
-  NETWORK_NAME="$(yq -r '.HOST.PODMAN_NETWORK_NAME' ${CONFIG_YAML})"
+  setDefaultEnvVars
+  SERVICE_NAME="caddy"
   CADDYFILE="$(yq -r '.CADDY.CADDYFILE' ${CONFIG_YAML})"
   CADDY_PROXY_DOMAIN="$(yq -r '.CADDY.PROXY_DOMAIN' ${CONFIG_YAML})"
-  CADDY_ROOT_DIR="$(yq -r '.CADDY.ROOT_DIR' ${CONFIG_YAML})"
-  NEXTCLOUD_ROOT_DIR="$(yq -r '.NEXTCLOUD.ROOT_DIR' ${CONFIG_YAML})"
+  DATA_DIR="$(yq -r '.CADDY.DATA_DIR' ${CONFIG_YAML})"
+  NEXTCLOUD_DATA_DIR="$(yq -r '.NEXTCLOUD.DATA_DIR' ${CONFIG_YAML})"
   SERVER_IP=$(hostname -I | awk '{print $1}')
   NEXTCLOUD_HTTP_PORT="$(yq -r '.NEXTCLOUD.HTTP_PORT' ${CONFIG_YAML})"
   NEXTCLOUD_HTTPS_PORT="$(yq -r '.NEXTCLOUD.HTTPS_PORT' ${CONFIG_YAML})"
   VAULTWARDEN_HTTPS_PORT="$(yq -r '.VAULTWARDEN.HTTPS_PORT' ${CONFIG_YAML})"
   VAULTWARDEN_ROCKET_PORT="$(yq -r '.VAULTWARDEN.ROCKET_PORT' ${CONFIG_YAML})"
   PODMAN_AUTO_UPDATE_STRATEGY="$(yq -r '.HOST.PODMAN_AUTO_UPDATE_STRATEGY' ${CONFIG_YAML})"
+  CONTAINER_IMAGE="$(yq -r '.CADDY.CONTAINER_IMAGE' "${CONFIG_YAML}")" 
+  START_ON_BOOT="$(yq -r '.CADDY.START_ON_BOOT' "${CONFIG_YAML}")" 
+  INTERNAL_TLS="$(yq -r '.CADDY.INTERNAL_TLS' "${CONFIG_YAML}")" 
 }
 
-
-function showStatus() {
-  SERVICES=(
-    caddy
-  )
-  for  i in ${!SERVICES[@]}; do
-        echo "###################################################"
-        echo "Show status for service ${SERVICES[$i]}"
-        ${SYSTEMCTL_CMD} --no-pager is-active  ${SERVICES[$i]}
-  done
-}
 
 function printEnvVars() {
-  echo CONFIG_YAML=${CONFIG_YAML}
-  echo QUADLET_DIR=${QUADLET_DIR}
-  echo SYSTEMD_UNIT_DIR=${SYSTEMD_UNIT_DIR}
-  echo SYSTEMCTL_CMD=${SYSTEMCTL_CMD}
-  echo NETWORK_NAME=${NETWORK_NAME}
+  printDefaultEnvVars
   echo PODMAN_AUTO_UPDATE_STRATEGY=${PODMAN_AUTO_UPDATE_STRATEGY}
   echo SERVER_IP=${SERVER_IP}
   echo CADDY_PROXY_DOMAIN=${CADDY_PROXY_DOMAIN}
   echo CADDYFILE=${CADDYFILE}
-  echo CADDY_ROOT_DIR=${CADDY_ROOT_DIR}
-  echo NEXTCLOUD_ROOT_DIR=${NEXTCLOUD_ROOT_DIR}
+  echo DATA_DIR=${DATA_DIR}
+  echo NEXTCLOUD_DATA_DIR=${NEXTCLOUD_DATA_DIR}
   echo NEXTCLOUD_HTTP_PORT=${NEXTCLOUD_HTTP_PORT}
   echo NEXTCLOUD_HTTPS_PORT=${NEXTCLOUD_HTTPS_PORT}
   echo VAULTWARDEN_HTTPS_PORT=${VAULTWARDEN_HTTPS_PORT}
   echo VAULTWARDEN_ROCKET_PORT=${VAULTWARDEN_ROCKET_PORT}
+  echo CONTAINER_IMAGE=${CONTAINER_IMAGE}
+  echo START_ON_BOOT=${START_ON_BOOT}
+  echo INTERNAL_TLS=${INTERNAL_TLS}
 }
 
-function installCaddy() {
-  CreatePodmanNetwork
-  CreateQuadletCaddy
-  postInstall
-  showStatus
-}
-
-
-function usage() {
-  echo "##################"
-  echo "Parameters available"
-  echo "-c <path-to-config-file> (required) "
-  echo "-i to install Nextcloud"
-  echo "-u to uninstall Nextcloud"
-  echo "-s to show status of Nextcloud services"
-}
-
-function checkpCLIParams() {
-  while getopts "iusc:" OPTNAME; do
-    case "${OPTNAME}" in
-      i )
-        echo "Runmode Option ${OPTNAME} is specified"
-        RUN_MODE="INSTALL"
-        ;;
-      u )
-        echo "Runmode Option ${OPTNAME} is specified"
-        RUN_MODE="UNINSTALL"
-        ;;
-      s )
-        echo "Runmode Option ${OPTNAME} is specified"
-        RUN_MODE="STATUS"
-        ;;
-      c )
-        echo "config file used is ${OPTARG} is specified"
-        CONFIG_YAML=${OPTARG}
-        ;;
-      * )
-        echo "unknown parameter specified"
-        usage
-        exit 1
-        ;;
-    esac
-  done
-  if [ $OPTIND -eq 1 ]; then 
-    echo "No options were passed"; 
-    usage
-    exit 1
-  fi
-
-  if [ -z "${CONFIG_YAML+x}" ]  || [ ! -f ${CONFIG_YAML} ]; then 
-    echo "Config file does not exist Please specify an existing config file witch -c"; 
-    usage
-  fi
-  if [ -z "${RUN_MODE+x}" ]; then
-     echo "No Runmode mode specified specify either -c or -u"
-     usage
-     exit 1
-  fi   
-}
-
-# main start here
-checkpCLIParams $*
-setEnvVars
-printEnvVars
-case "${RUN_MODE}" in 
-   "INSTALL" )
-     echo "installing"
-     installCaddy ;;
-   "UNINSTALL")
-     echo "uninstalling"
-     uninstallCaddy ;;
-   "STATUS")
-     echo "showing status"
-     showStatus ;;
-   * )
-     echo "Invalid Installation mode specified specifed us either -c or -u parameter"
-     usage; 
-     exit 1
-     ;;
- esac    
-
+main "$@"
